Django_Dashboard/requirements.txt
# Django dashboard requirements
# Install: pip install -r requirements.txt

django==5.1.3
pymongo==4.10.1
gunicorn==23.0.0
django-cors-headers==4.6.0
prometheus-client==0.21.0
PyJWT==2.9.0
bcrypt==4.2.0

# Async task queue (for triggering rebuilds)
celery==5.4.0
redis==5.1.1
